find neighboorhoods now takes a float input as the multiple of diameters entry and fixing issue where the search was looking "periodicially" on opposite edges

+++ Source/DREAM3DLib/StatisticsFilters/FindNeighborhoods.h
@@ -74,7 +74,7 @@ class DREAM3DLib_EXPORT FindNeighborhoods : public AbstractFilter
     DREAM3D_INSTANCE_STRING_PROPERTY(NeighborhoodListArrayName)
     DREAM3D_INSTANCE_STRING_PROPERTY(NumNeighborsArrayName)
 
-    DREAM3D_INSTANCE_PROPERTY(int, MultiplesOfAverage)
+    DREAM3D_INSTANCE_PROPERTY(float, MultiplesOfAverage)
 
     virtual const std::string getGroupName() { return DREAM3D::FilterGroups::StatisticsFilters; }
     virtual const std::string getSubGroupName() { return DREAM3D::FilterSubGroups::MorphologicalFilters; }
